Ignore .rvmrc
Currently, Pow! doesn't support .ruby-version for setting the current
ruby and uses the system ruby. Because of this, I have to manually
create .rvmrc, which is respected, to properly set that.

There is currently an issue for Pow! to fix this, but it has been
blocked for six months. I may take a whack at it later this week to
allow me to undo this travesty.

https://github.com/37signals/pow/pull/327